Refactor HomePage.jsx into smaller, maintainable components following App.jsx pattern
